+++ src/Logitar.Portal.Contracts/Constants/Application.cs
@@ -1,0 +1,6 @@
+﻿namespace Logitar.Portal.Contracts.Constants;
+
+public static class Application
+{
+  public static readonly Version Version = new(2, 0, 0);
+}