+++ src/Logitar.Portal.Contracts/Constants/Headers.cs
@@ -1,0 +1,6 @@
+﻿namespace Logitar.Portal.Contracts.Constants;
+
+public static class Headers
+{
+  public const string Authorization = nameof(Authorization);
+}