+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Reimplemented HTTP client.
 - Enhanced realm JWT secret.
 - Reimplemented password recovery.
+- Implemented Basic authentication.
 
 ## [1.1.5] - 2022-10-27